Backend-AI/requirements.txt

# https://pytorch.org/get-started/previous-versions/
# heroku config:set TORCH_CUDA_ARCH_LIST=""
# heroku config:set FORCE_CPU=1
# -f https://download.pytorch.org/whl/cpu
--index-url https://download.pytorch.org/whl/cpu
--extra-index-url https://pypi.org/simple
torch==2.4.0+cpu
torchvision==0.19.0+cpu

annotated-types==0.7.0
anyio==4.4.0
asyncio==3.4.3
bcrypt==4.2.0
certifi==2024.8.30
cffi==1.17.0
click==8.1.7
colorama==0.4.6
cryptography==43.0.0
dnspython==2.6.1
ecdsa==0.19.0
exceptiongroup==1.2.2
fastapi==0.112.0
filelock==3.15.4
fsspec==2024.6.1
greenlet==3.0.3
h11==0.14.0
idna==3.7
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==2.1.0
passlib==1.7.4
pillow==10.4.0
psycopg2==2.9.9
pyasn1==0.6.0
pycparser==2.22
pydantic==2.8.2
pydantic-settings==2.4.0
pydantic_core==2.20.1
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
rsa==4.9
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
sympy==1.13.2
typing_extensions==4.12.2
uvicorn==0.30.6
# gunicorn
# shutil